requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
PyJWT>=2.8.0
python-dotenv>=1.0.0
slowapi>=0.1.9
pydantic>=2.0
